mgr/vol: allow disabling purging of subvolumes from trash
Add a new MGR config option that allows subvolume to be moved to trash
but not to be deleted/purged from trash.

Fixes: https://tracker.ceph.com/issues/61903

+++ src/pybind/mgr/volumes/fs/operations/trash.py
@@ -51,7 +51,7 @@ class Trash(GroupTemplate):
         """
         return self._get_single_dir_entry(exclude_list)
 
-    def purge(self, trashpath, should_cancel):
+    def purge(self, trashpath, should_purge, should_cancel):
         """
         purge a trash entry.
 
@@ -61,24 +61,30 @@ class Trash(GroupTemplate):
         """
         def rmtree(root_path):
             log.debug("rmtree {0}".format(root_path))
+
             try:
                 with self.fs.opendir(root_path) as dir_handle:
                     d = self.fs.readdir(dir_handle)
-                    while d and not should_cancel():
+                    while d and should_purge() and not should_cancel():
                         if d.d_name not in (b".", b".."):
                             d_full = os.path.join(root_path, d.d_name)
                             if d.is_dir():
                                 rmtree(d_full)
                             else:
-                                self.fs.unlink(d_full)
+                                # check purge config option again, its value
+                                # may change while this loop is on and in such
+                                # case we want purge to halt immediately.
+                                if should_purge():
+                                    self.fs.unlink(d_full)
                         d = self.fs.readdir(dir_handle)
             except cephfs.ObjectNotFound:
                 return
             except cephfs.Error as e:
                 raise VolumeException(-e.args[0], e.args[1])
+
             # remove the directory only if we were not asked to cancel
             # (else we would fail to remove this anyway)
-            if not should_cancel():
+            if should_purge() and not should_cancel():
                 self.fs.rmdir(root_path)
 
         # catch any unlink errors
